# src/docs/swagger.yaml
definitions:
...
  model.Movies:
    properties:
      bannerPicURL:
        type: string
      coverPicURL:
        type: string
      description:
        type: string
      fsk:
        type: integer
      id:
        type: string
      rating:
        type: number
      releaseDate:
        type: string
      timeInMin:
        type: integer
      title:
        type: string
      trailerURL:
        type: string
    type: object
  model.Orders:
    properties:
      id:
        type: string
      isPaid:
        type: boolean
      paymentMethodID:
        type: string
      totalprice:
        type: integer
      userID:
        type: string
    type: object
...
  model.Seats:
    properties:
      cinemaHallID:
        type: string
      columnNr:
        type: integer
      id:
        type: string
      rowNr:
        type: integer
      seatCategoryID:
        type: string
      type:
        type: string
    type: object
  model.Tickets:
    properties:
      eventSeatID:
        type: string
      id:
        type: string
      orderID:
        type: string
      price:
        type: integer
      priceCategoryID:
        type: string
      validated:
        type: boolean
    type: object
...
  models.CreateOrderDTO:
    properties:
      eventSeatPriceCategories:
        items:
          properties:
            eventSeatId:
              type: string
            priceCategoryId:
              type: string
          type: object
        type: array
      paymentMethodID:
        type: string
    type: object
...
  models.GetOrderDTO:
    properties:
      order:
        $ref: '#/definitions/model.Orders'
      tickets:
        items:
          properties:
            seat:
              $ref: '#/definitions/model.Seats'
            ticket:
              $ref: '#/definitions/model.Tickets'
          type: object
        type: array
    type: object
...
  models.IdResponse:
    properties:
      id:
        type: string
    type: object
  models.KTSErrorMessage:
    properties:
      errorMessage:
        type: string
    type: object
...
  models.MovieDTOCreate:
    properties:
      actorsID:
        items:
          properties:
            id:
              type: string
          type: object
        type: array
      bannerPicURL:
        type: string
      coverPicURL:
        type: string
      description:
        type: string
      fsk:
        type: integer
      genresID:
        items:
          properties:
            id:
              type: string
          type: object
        type: array
      id:
        type: string
      producersID:
        items:
          properties:
            id:
              type: string
          type: object
        type: array
      rating:
        type: number
      releaseDate:
        type: string
      timeInMin:
        type: integer
      title:
        type: string
      trailerURL:
        type: string
    type: object
  models.MovieWithEverything:
    properties:
      actors:
        items:
          properties:
            birthdate:
              type: string
            description:
              type: string
            id:
              type: string
            name:
              type: string
          type: object
        type: array
      bannerPicURL:
        type: string
      coverPicURL:
        type: string
      description:
        type: string
      fsk:
        type: integer
      genres:
        items:
          properties:
            genreName:
              type: string
            id:
              type: string
          type: object
        type: array
      id:
        type: string
      producers:
        items:
          properties:
            birthdate:
              type: string
            description:
              type: string
            id:
              type: string
            name:
              type: string
          type: object
        type: array
      rating:
        type: number
      releaseDate:
        type: string
      reviews:
        items:
          properties:
            comment:
              type: string
            datetime:
              type: string
            id:
              type: string
            isSpoiler:
              type: boolean
            movieID:
              type: string
            rating:
              type: integer
            userID:
              type: string
          type: object
        type: array
      timeInMin:
        type: integer
      title:
        type: string
      trailerURL:
        type: string
    type: object
  models.MovieWithGenres:
    properties:
      bannerPicURL:
        type: string
      coverPicURL:
        type: string
      description:
        type: string
      fsk:
        type: integer
      genres:
        items:
          properties:
            genreName:
              type: string
            id:
              type: string
          type: object
        type: array
      id:
        type: string
      rating:
        type: number
      releaseDate:
        type: string
      timeInMin:
        type: integer
      title:
        type: string
      trailerURL:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /events/{eventId}/book:
    post:
      consumes:
      - application/json
      description: Create order
      parameters:
      - description: Event ID
        in: path
        name: eventId
        required: true
        type: string
      - description: Order data
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/models.CreateOrderDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
      summary: Create order
      tags:
      - Orders
  /events/{eventId}/reserve:
    post:
      consumes:
      - application/json
      description: Create order
      parameters:
      - description: Event ID
        in: path
        name: eventId
        required: true
        type: string
      - description: Order data
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/models.CreateOrderDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
      summary: Create order
      tags:
      - Orders
...
  /movies:
    get:
      consumes:
      - application/json
      description: Get movies
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Movies'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
      summary: Get movies
      tags:
      - Movies
    post:
      consumes:
      - application/json
      description: Create movie
      parameters:
      - description: Movie data
        in: body
        name: movie
        required: true
        schema:
          $ref: '#/definitions/models.MovieDTOCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
      summary: Create movie
      tags:
      - Movies
  /movies/{id}:
    get:
      consumes:
      - application/json
      description: Get Movie By Id
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MovieWithEverything'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
      summary: Get Movie By Id
      tags:
      - Movies
...
  /movies/genres:
    get:
      consumes:
      - application/json
      description: Get movies with genres
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.MovieWithGenres'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
      summary: Get movies with genres
      tags:
      - Movies
  /orders:
    get:
      consumes:
      - application/json
      description: Get orders for user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.GetOrderDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
      summary: Get orders
      tags:
      - Orders
  /orders/{orderId}:
    get:
      consumes:
      - application/json
      description: Get order by id
      parameters:
      - description: Order ID
        in: path
        name: orderId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetOrderDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
      summary: Get order by id
      tags:
      - Orders
...
swagger: "2.0"